Refactor asset data for consistency
